Prevented player from trading with self, and quick trade.

+++ TradeRequest/Commands.cs
@@ -1,8 +1,5 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using Terraria;
 using TShockAPI;
 
@@ -24,59 +21,70 @@ namespace TradeRequest
             }
 
             var player = args.Player;
+            int amount = 0;
             if (player.TPlayer.trashItem.netID == 0)
             {
                 player.SendErrorMessage("You must have an item in the Trash Bin to trade.");
                 return;
             }
+
             if (args.Parameters[0] == "confirm")
             {
                 TradePlayer trader = TradeRequest.TradeActive[player.Index];
                 TradePlayer target = TradeRequest.TradeActive[trader.target];
-                TSPlayer targetplayer = TShock.Players[trader.target];
-                if (trader.active && trader.item == player.TPlayer.trashItem.netID && target.active && target.item == targetplayer.TPlayer.trashItem.netID)
+                if ((DateTime.UtcNow - trader.confirmlock).TotalSeconds > 3)
                 {
-                    trader.active = false;
-                    target.active = false;
-                    var pItem = trader.item;
-                    var pStack = player.TPlayer.trashItem.stack;
-                    var pPrefix = player.TPlayer.trashItem.prefix;
-                    var tItem = target.item;
-                    var tStack = targetplayer.TPlayer.trashItem.stack;
-                    var tPrefix = targetplayer.TPlayer.trashItem.prefix;
-
-                    player.TPlayer.trashItem.netDefaults(tItem);
-                    if (player.TPlayer.trashItem.netID != 0)
+                    #region Trade Section
+                    TSPlayer targetplayer = TShock.Players[trader.target];
+                    if (trader.active && target.active)
                     {
-                        player.TPlayer.trashItem.stack = tStack;
-                        player.TPlayer.trashItem.prefix = tPrefix;
-                    }
+                        trader.active = false;
+                        target.active = false;
+                        var pItem = trader.item;
+                        var pStack = player.TPlayer.trashItem.stack;
+                        var pPrefix = player.TPlayer.trashItem.prefix;
+                        var pAmount = trader.amount;
+                        var tItem = target.item;
+                        var tStack = targetplayer.TPlayer.trashItem.stack;
+                        var tPrefix = targetplayer.TPlayer.trashItem.prefix;
+                        var tAmout = target.amount;
 
-                    targetplayer.TPlayer.trashItem.netDefaults(pItem);
-                    if (targetplayer.TPlayer.trashItem.netID != 0)
+                        player.TPlayer.trashItem.netDefaults(tItem);
+                        if (player.TPlayer.trashItem.netID != 0)
+                        {
+                            player.TPlayer.trashItem.stack = tStack;
+                            player.TPlayer.trashItem.prefix = tPrefix;
+                        }
+
+                        targetplayer.TPlayer.trashItem.netDefaults(pItem);
+                        if (targetplayer.TPlayer.trashItem.netID != 0)
+                        {
+                            targetplayer.TPlayer.trashItem.stack = pStack;
+                            targetplayer.TPlayer.trashItem.prefix = pPrefix;
+                        }
+
+                        NetMessage.SendData(5, -1, -1, Main.player[player.Index].trashItem.name, player.Index, 179, (float)tPrefix);
+                        NetMessage.SendData(5, player.Index, -1, Main.player[player.Index].trashItem.name, player.Index, 179, (float)tPrefix);
+                        NetMessage.SendData(5, -1, -1, Main.player[targetplayer.Index].trashItem.name, targetplayer.Index, 179, (float)pPrefix);
+                        NetMessage.SendData(5, targetplayer.Index, -1, Main.player[targetplayer.Index].trashItem.name, targetplayer.Index, 179, (float)pPrefix);
+                        TSPlayer.All.SendInfoMessage("{0} ([i/s{1}:{2}]) traded successfully with {3} ([i/s{4}:{5}]).", player.Name, pStack, pItem, targetplayer.Name, tStack, tItem);
+                        TShock.Log.ConsoleInfo("TRADEINFO: {0} traded x{1} {2} for {3}'s x{4} {5}", player.Name, pStack, Main.player[targetplayer.Index].trashItem.name, targetplayer.Name, tStack, Main.player[player.Index].trashItem.name);
+                        return;
+                    }
+                    else if ((trader.active && trader.item != player.TPlayer.trashItem.netID) || (target.active && target.item != targetplayer.TPlayer.trashItem.netID))
                     {
-                        targetplayer.TPlayer.trashItem.stack = pStack;
-                        targetplayer.TPlayer.trashItem.prefix = pPrefix;
+                        player.SendErrorMessage("Trade Failed. Item is no longer in the Trash Bin or Scamming.");
+                        return;
                     }
-
-                    NetMessage.SendData(5, -1, -1, Main.player[player.Index].trashItem.name, player.Index, 179, (float)tPrefix);
-                    NetMessage.SendData(5, player.Index, -1, Main.player[player.Index].trashItem.name, player.Index, 179, (float)tPrefix);
-                    NetMessage.SendData(5, -1, -1, Main.player[targetplayer.Index].trashItem.name, targetplayer.Index, 179, (float)pPrefix);
-                    NetMessage.SendData(5, targetplayer.Index, -1, Main.player[targetplayer.Index].trashItem.name, targetplayer.Index, 179, (float)pPrefix);
-                    TSPlayer.All.SendInfoMessage("{0}([i/s{1}:{2}]) traded successfully with {3}([i/s{4}:{5}]).", player.Name, pStack, pItem, targetplayer.Name, tStack, tItem);
-                    TShock.Log.ConsoleInfo("TRADEINFO: {0} traded x{1} {2} for {3}'s x{4} {5}", player.Name, pStack, Main.player[targetplayer.Index].trashItem.name, targetplayer.Name, tStack, Main.player[player.Index].trashItem.name);
-                    return;
-                }
-                else if ((trader.active && trader.item != player.TPlayer.trashItem.netID) || (target.active && target.item != targetplayer.TPlayer.trashItem.netID))
-                {
-                    player.SendErrorMessage("Trade Failed. Item is no longer in the Trash Bin or Scamming.");
-                    return;
+                    else
+                    {
+                        player.SendInfoMessage("You must send a trade request. Please use /trade <player>");
+                        return;
+                    }
+                    #endregion
                 }
                 else
-                {
-                    player.SendInfoMessage("You must send a trade request. Please use /trade <player>");
-                    return;
-                }
+                    args.Player.SendErrorMessage("You must wait a few seconds before finalizing the trade.");
             }
 
             var foundplr = TShock.Utils.FindPlayer(args.Parameters[0]);
@@ -93,6 +101,12 @@ namespace TradeRequest
             else
             {
                 var found = foundplr[0];
+                if (found.Index == args.Player.Index)
+                {
+                    args.Player.SendErrorMessage("You cannot trade with yourself.");
+                    return;
+                }
+
                 TradePlayer trader = TradeRequest.TradeActive[player.Index];
                 if (!trader.active)
                 {
@@ -104,6 +118,8 @@ namespace TradeRequest
                     if (target.active == true)
                     {
                         target.time = DateTime.UtcNow;
+                        trader.confirmlock = DateTime.UtcNow;
+                        target.confirmlock = DateTime.UtcNow;
                     }
                 }
                 else
@@ -111,8 +127,8 @@ namespace TradeRequest
                     player.SendErrorMessage("You already have a trade active.");
                     return;
                 }
-                player.SendInfoMessage("You have sent a trade request to {0}. You can cancel your trade by removing the item.", found.Name);
-                found.SendInfoMessage("{0} would like to trade [i/s{1}:{2}]. Type [c/FFFFFF:/trade confirm] after [c/FFFFFF:/trade {0}].", player.Name, player.TPlayer.trashItem.stack, player.TPlayer.trashItem.netID);
+                player.SendInfoMessage("You have sent a trade request to {0}. Cancel the trade by removing the item.", found.Name);
+                found.SendInfoMessage("{0} would like to trade {1}. Type [c/FFFFFF:/trade confirm] after [c/FFFFFF:/trade {0}].", player.Name, trader.amount > 0 ? trader.amount.ToString() + " Gem(s)" : "[i/s" + player.TPlayer.trashItem.stack + ":" + player.TPlayer.trashItem.netID + "]");
             }
         }
     }
